reverted modal, indicator changes. reorganized code
