+++ client/src/Core/Components/MainLayout.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Header } from './Header';
-import { styled, Spinner, SpinnerSize, FontSizes, FontWeights, Text } from '@fluentui/react';
+import { styled, Spinner, SpinnerSize, FontSizes, FontWeights, Text, Separator, NavBase } from '@fluentui/react';
 import { SimpleComponentStyles, IThemeOnlyProps, IStylesOnly } from '../Utils/FluentUI/typings.fluent-ui';
 import { PagesRouter } from '../../Router/PagesRouter';
 import { useObserver } from 'mobx-react-lite';
@@ -11,9 +11,11 @@ import { StudentPage } from '../../Features/StudentView/StudentPage';
 import learnLogo from '../../Assets/icon_learn_062020.png';
 import { useQueryValue } from '../Hooks';
 import { ErrorPage } from './ErrorsPage';
-import { NavPivot } from './Navbar/NavPivot';
+import { PublishControlArea } from '../../Features/PublishAssignment/PublishControlArea';
+import { PublishSuccessMessageBar } from '../../Features/PublishAssignment/PublishSuccessMessageBar';
+import * as NavBarBase from './Navbar'
 
-type MainLayoutStyles = SimpleComponentStyles<'root' | 'spinner' | 'content' | 'assignmentTitle'>;
+type MainLayoutStyles = SimpleComponentStyles<'root' | 'spinner' | 'content' | 'assignmentTitle' | 'navAndControlArea' | 'separator'>;
 
 const MainLayoutInner = ({ styles }: IStylesOnly<MainLayoutStyles>): JSX.Element => {
   useAssignmentInitializer();
@@ -51,7 +53,12 @@ const MainLayoutInner = ({ styles }: IStylesOnly<MainLayoutStyles>): JSX.Element
               <Text variant="xLargePlus" className={classes.assignmentTitle}>
                 {assignmentStore.assignment.name}
               </Text>
-              <NavPivot/>
+              <div className={classes.navAndControlArea}>
+                <NavBarBase.NavbarTop />                  
+                <PublishControlArea/>
+              </div>
+              <Separator className={classes.separator} />
+              <PublishSuccessMessageBar isPublished={assignmentStore.assignment.publishStatus === 'Published'}/>
               <PagesRouter />
             </>
           ) : (
@@ -69,7 +76,8 @@ const mainLayoutStyle = ({ theme }: IThemeOnlyProps): MainLayoutStyles => {
       {
         height: '100vh',
         display: 'flex',
-        flexDirection: 'column'
+        flexDirection: 'column',
+        backgroundColor: theme.palette.neutralLighterAlt,
       }
     ],
     spinner: [
@@ -101,6 +109,24 @@ const mainLayoutStyle = ({ theme }: IThemeOnlyProps): MainLayoutStyles => {
         paddingLeft: `calc(${theme.spacing.l1}*1.6)`,
         paddingBottom: `calc(${theme.spacing.l1}*0.5)`,
         paddingTop:`calc(${theme.spacing.l1}*1.5)`,
+      }
+    ],
+    navAndControlArea: [
+      {
+        display: 'flex',
+        backgroundColor: theme.palette.neutralLighterAlt,
+        paddingLeft: `calc(${theme.spacing.l1} * 1.6)`,
+        paddingRight: `calc(${theme.spacing.l1} * 1.6)`,     
+        height:`calc(${theme.spacing.l1} * 2.7)`,
+        flexDirection: 'row',
+        justifyContent: 'space-between',  
+      }
+    ],
+    separator: [
+      {
+        height: '1px',
+        backgroundColor: '#E1DFDD',
+        padding: `0px 0px`
       }
     ]
   };
