+++ client/src/Core/Components/Navbar/NavbarTop.tsx
@@ -5,15 +5,13 @@ import { useHistory, useLocation } from 'react-router-dom';
 import _ from 'lodash';
 import { configurationRoutes, viewRoutes } from '../../../Router/RoutesConfiguration';
 import { useStore } from '../../../Stores/Core';
-import { styled, PivotItem, Pivot, IPivotStyles, MessageBarType, AnimationClassNames, mergeStyles, Separator, FontWeights } from '@fluentui/react';
+import { styled, PivotItem, Pivot, IPivotStyles, FontWeights } from '@fluentui/react';
 import { IThemeOnlyProps, IStylesOnly, SimpleComponentStyles } from '../../Utils/FluentUI/typings.fluent-ui';
 import { Assignment } from '../../../Models/Assignment.model';
-import { PublishControlArea } from '../../../Features/PublishAssignment/PublishControlArea';
 import { themedClassNames } from '../../Utils/FluentUI';
-import { AutohideMessageBar } from '../AutohideMessageBar';
 import { pagesDisplayNames } from '../../../Router/Consts';
 
-type NavWrapperStyles = SimpleComponentStyles<'root'  | 'navArea' | 'messageBar' | 'separator'>;
+type NavWrapperStyles = SimpleComponentStyles<'root'>;
 
 const getNavLinks = (assignment: Assignment): INavLink[] => [
   ...(assignment.publishStatus === 'Published' ? 
@@ -61,39 +59,23 @@ const NavPivotInner = ({ styles }: IStylesOnly<NavWrapperStyles>): JSX.Element |
 
       return (
         <div className={classes.root}>
-          <div className={classes.navArea}>
-            <Pivot
-              selectedKey={selectedNavKey}
-              onLinkClick={handleLinkClick}
-              styles={themedClassNames(navStyles)}
-            >
-              {_.map(getMappedLinks(assignmentStore.assignment), link=>{
-                let urlWithSearchParams = link.url + 
-                  (link.search!==undefined && link.search.length>0 ? '?'+link.search : '')
-                return(            
-                  <PivotItem
-                    headerText={link.name}
-                    itemIcon={link.icon}    
-                    itemKey={urlWithSearchParams} 
-                    headerButtonProps={{disabled:link.disabled?link.disabled: false}}
-                />)
-              })}
-            </Pivot>
-            <PublishControlArea />            
-          </div>
-          <Separator className={classes.separator} />
-          {assignmentStore.assignment && (
-            <AutohideMessageBar
-              messageBarType={MessageBarType.success}
-              isMultiline={false}
-              className={classes.messageBar}
-              showMessageBar={
-                assignmentStore.assignment.publishStatus === 'Published' && assignmentStore.isChangingPublishState === false
-              }
-            >
-              Your assignment was published successfully
-            </AutohideMessageBar>
-          )}
+          <Pivot
+            selectedKey={selectedNavKey}
+            onLinkClick={handleLinkClick}
+            styles={themedClassNames(navStyles)}
+          >
+            {_.map(getMappedLinks(assignmentStore.assignment), link=>{
+              let urlWithSearchParams = link.url + 
+                (link.search!==undefined && link.search.length>0 ? '?'+link.search : '')
+              return(            
+                <PivotItem
+                  headerText={link.name}
+                  itemIcon={link.icon}    
+                  itemKey={urlWithSearchParams} 
+                  headerButtonProps={{disabled:link.disabled?link.disabled: false}}
+              />)
+            })}
+          </Pivot>
         </div>
       )
     }
@@ -103,34 +85,10 @@ const NavPivotInner = ({ styles }: IStylesOnly<NavWrapperStyles>): JSX.Element |
 const NavWrapperStyles = ({ theme }: IThemeOnlyProps): NavWrapperStyles => ({
   root: [
     {
-      display:'block',
-      backgroundColor: theme.palette.neutralLighterAlt,
-      paddingLeft: `calc(${theme.spacing.l1} * 1.6)`,
-      paddingRight: `calc(${theme.spacing.l1} * 1.6)`,   
-
-    }
-  ],
-  navArea:[
-    {
       display: 'flex',
       height:`calc(${theme.spacing.l1} * 2.7)`,
       flexDirection: 'row',
       justifyContent: 'space-between',  
-    }
-  ],
-  messageBar: [
-    mergeStyles(AnimationClassNames.fadeIn200, {
-      height: theme.spacing.l2,
-      width: 'auto',
-      marginLeft: 'auto',
-      marginTop: `calc(${theme.spacing.s1}*1.6)`
-    })
-  ],
-  separator: [
-    {
-      height: '1px',
-      backgroundColor: '#E1DFDD',
-      padding: `0px 0px`
     }
   ]
 });
@@ -158,4 +116,4 @@ const navStyles = ({ theme }: IThemeOnlyProps): Partial<IPivotStyles> => ({
   }
 })
 
-export const NavPivot = styled(NavPivotInner, NavWrapperStyles);
+export const NavbarTop = styled(NavPivotInner, NavWrapperStyles);
