+++ client/src/Core/Components/Navbar/index.ts
@@ -1,0 +1,2 @@
+export * from './NavbarTop';
+export * from './NavbarLHP';
