+++ client/src/Features/PublishAssignment/PublishSuccessMessageBar.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect, useRef } from 'react';
 import { IThemeOnlyProps, IStylesOnly } from '../../Core/Utils/FluentUI/typings.fluent-ui';
-import { styled, MessageBar, MessageBarType, IMessageBarStyles } from '@fluentui/react';
+import { styled, MessageBar, MessageBarType, IMessageBarStyles, mergeStyles, AnimationClassNames } from '@fluentui/react';
 import { themedClassNames } from '../../Core/Utils/FluentUI';
 
 interface PublishSuccessMessageBarProps {
@@ -45,7 +45,11 @@ const PublishSuccessMessageBarInner = ({
 
   if (isShown) {
     return (
-      <MessageBar messageBarType={MessageBarType.success} isMultiline={false} styles={themedClassNames(styles)}>
+      <MessageBar 
+        messageBarType={MessageBarType.success} 
+        isMultiline={false} 
+        styles={themedClassNames(styles)}
+      >
         Your assignment was published successfully
       </MessageBar>
     );
@@ -55,11 +59,15 @@ const PublishSuccessMessageBarInner = ({
 
 const publishSuccessMessageBarStyles = ({ theme: _theme }: IThemeOnlyProps): PublishSuccessMessageBarStyles => ({
   root: [
-    {
+    mergeStyles(AnimationClassNames.fadeIn200, {
+      height: _theme.spacing.l2,
       width: 'auto',
-      marginLeft: 'auto'
-    }
+      marginLeft: '32px',
+      marginRight: '32px',
+      marginTop: `calc(${_theme.spacing.s1}*1.6)`
+    })
   ]
+  
 });
 
 export const PublishSuccessMessageBar = styled(PublishSuccessMessageBarInner, publishSuccessMessageBarStyles);
